Move the main loop into it's own module
