+++ src/lib.rs
@@ -43,11 +43,12 @@ macro_rules! try_msg_or {
     }
 }
 
-pub mod rvi;
-pub mod sota_dbus;
-pub mod jsonrpc;
-
+pub mod main_loop;
 pub mod configuration;
-pub mod handler;
-pub mod message;
+
+mod rvi;
+mod sota_dbus;
+mod jsonrpc;
+mod handler;
+mod message;
 mod persistence;
